Use u32 instead of size_t for now

+++ include/secret_base.h
@@ -25,6 +25,6 @@ void TrySetCurSecretBaseIndex(void);
 void CheckPlayerHasSecretBase(void);
 void ToggleSecretBaseEntranceMetatile(void);
 void ScriptContext_Enable(void);
-void ReceiveSecretBasesData(void *records, size_t recordSize, u8 linkIdx);
+void ReceiveSecretBasesData(void *records, u32 recordSize, u8 linkIdx);
 
 #endif //GUARD_SECRET_BASE_H
